+++ packages/web-data/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@imgly/background-removal-data",
-  "version": "1.5.3",
+  "version": "1.5.5",
   "description": "Background Removal Data",
   "keywords": [
     "background-removal",
