+++ packages/web/changelog/1.5.0/20242503170808-Added_isnet_model_for_webgpu_Added.yaml
