+++ packages/web/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@imgly/background-removal",
-  "version": "1.5.3",
+  "version": "1.5.5",
   "description": "Background Removal in the Browser",
   "keywords": [
     "background-removal",
